Modify structure of historical_prices_tickers.json
Not all securities are available for all sources. It was solved by setting key:value pair as
'source':'-' when the source did not include that security. Not elegant and add code to
filter that in list comprehension.

Solution: Add a level in the json structure. Each security has a 'tickers':dict pair, wich
includes one 'source':'ticker' pair for all source with such security.