Bug fix:
Problem: in some condition, no bar opens in one or several of the momzones, which
leads to a missing count in count_fading_gaps. That further leads to an key error in the
the section lower where count_fading_gaps[xxx][momzone_label] is called.

To fix that, added a check for missing momzone labels in count_fading_gaps, and add a count of zero for any missing.